--- docker/mage/ashraf-magic/data_loaders/load_titanic.py
@@ -1,27 +1,0 @@
-import io
-import pandas as pd
-import requests
-from pandas import DataFrame
-
-if 'data_loader' not in globals():
-    from mage_ai.data_preparation.decorators import data_loader
-if 'test' not in globals():
-    from mage_ai.data_preparation.decorators import test
-
-
-@data_loader
-def load_data_from_api(**kwargs) -> DataFrame:
-    """
-    Template for loading data from API
-    """
-    url = 'https://raw.githubusercontent.com/datasciencedojo/datasets/master/titanic.csv?raw=True'
-
-    return pd.read_csv(url)
-
-
-@test
-def test_output(df) -> None:
-    """
-    Template code for testing the output of the block.
-    """
-    assert df is not None, 'The output is undefined'
