+++ docker/mage/ashraf-magic/pipelines/batch_pipeline/triggers.yaml
@@ -1,0 +1,12 @@
+triggers:
+- envs: []
+  last_enabled_at: 2024-03-28 16:19:55.483035
+  name: gcs_to_bigquery_trigger
+  pipeline_uuid: batch_pipeline
+  schedule_interval: null
+  schedule_type: api
+  settings: null
+  sla: null
+  start_time: 2024-03-28 16:19:00
+  status: active
+  variables: {}
