+++ docker/mage/mage_data/ashraf-magic/.cache/pipeline_details_mapping
@@ -1,1 +1,1 @@
-{"groups": {"type": {"python": ["ashraf-magic:example_pipeline", "ashraf-magic:batch_pipeline"]}}, "models": {"ashraf-magic:example_pipeline": {"added_at": null, "pipeline": {"created_at": null, "description": null, "name": "example_pipeline", "tags": null, "type": "python", "uuid": "example_pipeline", "repo_path": "ashraf-magic", "blocks": [{"downstream_blocks": ["fill_in_missing_values"], "name": "load_titanic", "type": "data_loader", "upstream_blocks": [], "uuid": "load_titanic"}, {"downstream_blocks": ["export_titanic_clean"], "name": "fill_in_missing_values", "type": "transformer", "upstream_blocks": ["load_titanic"], "uuid": "fill_in_missing_values"}, {"downstream_blocks": [], "name": "export_titanic_clean", "type": "data_exporter", "upstream_blocks": ["fill_in_missing_values"], "uuid": "export_titanic_clean"}]}, "updated_at": 1711637853.497939}, "ashraf-magic:batch_pipeline": {"added_at": null, "pipeline": {"created_at": "2024-03-28 14:58:23.122970+00:00", "description": null, "name": "batch_pipeline", "tags": [], "type": "python", "uuid": "batch_pipeline", "blocks": [{"downstream_blocks": ["gcs_to_bigquery"], "language": "python", "name": "gcs_loader", "type": "data_loader", "upstream_blocks": [], "uuid": "gcs_loader"}, {"downstream_blocks": [], "language": "python", "name": "gcs_to_bigquery", "type": "data_exporter", "upstream_blocks": ["gcs_loader"], "uuid": "gcs_to_bigquery"}]}, "updated_at": 1711639291.675504}}}
+{"groups": {"type": {"python": ["ashraf-magic:example_pipeline", "ashraf-magic:batch_pipeline"]}}, "models": {"ashraf-magic:example_pipeline": {"added_at": null, "pipeline": {"created_at": null, "description": null, "name": "example_pipeline", "tags": null, "type": "python", "uuid": "example_pipeline", "repo_path": "ashraf-magic", "blocks": [{"downstream_blocks": ["fill_in_missing_values"], "name": "load_titanic", "type": "data_loader", "upstream_blocks": [], "uuid": "load_titanic"}, {"downstream_blocks": ["export_titanic_clean"], "name": "fill_in_missing_values", "type": "transformer", "upstream_blocks": ["load_titanic"], "uuid": "fill_in_missing_values"}, {"downstream_blocks": [], "name": "export_titanic_clean", "type": "data_exporter", "upstream_blocks": ["fill_in_missing_values"], "uuid": "export_titanic_clean"}]}, "updated_at": 1711637853.497939}, "ashraf-magic:batch_pipeline": {"added_at": null, "pipeline": {"created_at": "2024-03-28 14:58:23.122970+00:00", "description": null, "name": "batch_pipeline", "tags": [], "type": "python", "uuid": "batch_pipeline", "blocks": [{"downstream_blocks": ["gcs_to_bigquery_dim_time", "gcs_to_bigquery_dim_customer", "gcs_to_bigquery_dim_location", "gcs_to_bigquery_dim_delivery_person", "export_to_bigquery_fact_order"], "language": "python", "name": "gcs_loader", "type": "data_loader", "upstream_blocks": [], "uuid": "gcs_loader"}, {"downstream_blocks": [], "language": "python", "name": "export_to_bigquery_fact_order", "type": "data_exporter", "upstream_blocks": ["gcs_loader"], "uuid": "export_to_bigquery_fact_order"}, {"downstream_blocks": [], "language": "python", "name": "gcs_to_bigquery_dim_delivery_person", "type": "data_exporter", "upstream_blocks": ["gcs_loader"], "uuid": "gcs_to_bigquery_dim_delivery_person"}, {"downstream_blocks": [], "language": "python", "name": "gcs_to_bigquery_dim_location", "type": "data_exporter", "upstream_blocks": ["gcs_loader"], "uuid": "gcs_to_bigquery_dim_location"}, {"downstream_blocks": [], "language": "python", "name": "gcs_to_bigquery_dim_customer", "type": "data_exporter", "upstream_blocks": ["gcs_loader"], "uuid": "gcs_to_bigquery_dim_customer"}, {"downstream_blocks": [], "language": "python", "name": "gcs_to_bigquery_dim_time", "type": "data_exporter", "upstream_blocks": ["gcs_loader"], "uuid": "gcs_to_bigquery_dim_time"}]}, "updated_at": 1711642722.032621}}}
