+++ docker/mage/mage_data/ashraf-magic/pipelines/batch_pipeline/.logs/1/20240328T162258_317607/export_to_bigquery_fact_order.log
@@ -1,0 +1,9 @@
+2024-03-28T16:23:37 {"block_run_id": 2, "block_uuid": "export_to_bigquery_fact_order", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "Start executing block with BlockExecutor.", "timestamp": 1711643017.734629, "uuid": "d8a9a84948a64966bb200369b8878718"}
+2024-03-28T16:24:11 {"block_run_id": 2, "block_uuid": "export_to_bigquery_fact_order", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "BigQuery initialized", "timestamp": 1711643051.539964, "uuid": "d564300e4a7d411d9c1d5b5d5a320ce9"}
+2024-03-28T16:24:11 {"block_run_id": 2, "block_uuid": "export_to_bigquery_fact_order", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u2514\u2500 Connecting to BigQuery warehouse...", "timestamp": 1711643051.873534, "uuid": "39d85112990c4d2c9d9a3215b1fc721c"}
+2024-03-28T16:24:11 {"block_run_id": 2, "block_uuid": "export_to_bigquery_fact_order", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "DONE", "timestamp": 1711643051.874479, "uuid": "cd059d60770d4ea688bc8fc208dc7c08"}
+2024-03-28T16:24:12 {"block_run_id": 2, "block_uuid": "export_to_bigquery_fact_order", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "", "timestamp": 1711643052.139008, "uuid": "cedbc2054d1d462dbc95c9efcb87434c"}
+2024-03-28T16:24:12 {"block_run_id": 2, "block_uuid": "export_to_bigquery_fact_order", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u251c\u2500", "timestamp": 1711643052.139391, "uuid": "27f54876de9c4939b9db93f2e5ab97c7"}
+2024-03-28T16:24:12 {"block_run_id": 2, "block_uuid": "export_to_bigquery_fact_order", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u2514\u2500 Exporting data to table 'gothic-sylph-387906.customer_activity_dat.fact_order'...", "timestamp": 1711643052.139559, "uuid": "401eaf9c8b5b4ecbbaabb0798ed846f1"}
+2024-03-28T16:24:23 {"block_run_id": 2, "block_uuid": "export_to_bigquery_fact_order", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "DONE", "timestamp": 1711643063.957697, "uuid": "2bdc07a76bce426b882f88e54c3ea106"}
+2024-03-28T16:24:23 {"block_run_id": 2, "block_uuid": "export_to_bigquery_fact_order", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "Finish executing block with BlockExecutor.", "timestamp": 1711643063.964901, "uuid": "5c441de4004843f8a83c6fcdc5d746c7"}
