+++ docker/mage/mage_data/ashraf-magic/pipelines/batch_pipeline/.logs/1/20240328T162258_317607/gcs_loader.log
@@ -1,0 +1,19 @@
+2024-03-28T16:23:09 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "Start executing block with BlockExecutor.", "timestamp": 1711642989.241918, "uuid": "9432c184548341efb85ce33922d883af"}
+2024-03-28T16:23:20 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "GoogleCloudStorage initialized", "timestamp": 1711643000.888175, "uuid": "8257d9168cfb4a119f17986fe90a34e2"}
+2024-03-28T16:23:20 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "\u2514\u2500 Loading data frame from bucket 'customer-activity-data-terraform' at key 'transformed_data/customer_dimension/part-00000-3a8daa39-85da-4732-9037-46825fca65c8-c000.parquet'...", "timestamp": 1711643000.93632, "uuid": "13261d82a6bb4a8dbb3a88e7acedb667"}
+2024-03-28T16:23:23 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "DONE", "timestamp": 1711643003.302531, "uuid": "cf89b924c61b49a58a4c99dfe729db8c"}
+2024-03-28T16:23:23 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "GoogleCloudStorage initialized", "timestamp": 1711643003.315008, "uuid": "3406f4a0ac0b4fea86603065094a9de3"}
+2024-03-28T16:23:23 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "\u2514\u2500 Loading data frame from bucket 'customer-activity-data-terraform' at key 'transformed_data/delivery_person_dimension/part-00000-f2c31ceb-ac8c-42d8-8002-b69646801844-c000.parquet'...", "timestamp": 1711643003.356903, "uuid": "0ffd9e4551ea4ff6ab74d0b26861f887"}
+2024-03-28T16:23:25 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "DONE", "timestamp": 1711643005.393203, "uuid": "dc3a4941d15f47df9eb2afaa4a6f3260"}
+2024-03-28T16:23:25 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "GoogleCloudStorage initialized", "timestamp": 1711643005.408726, "uuid": "db9337118aec4fefbdc4670e49cd464c"}
+2024-03-28T16:23:25 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "\u2514\u2500 Loading data frame from bucket 'customer-activity-data-terraform' at key 'transformed_data/location_dimension/part-00000-15bf43de-6df9-4552-a21c-4e2a440f44b6-c000.parquet'...", "timestamp": 1711643005.46207, "uuid": "e163114037ad4367a07b971b2807ab9b"}
+2024-03-28T16:23:27 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "DONE", "timestamp": 1711643007.995521, "uuid": "dd8dd802f78c4809b794484679f3e997"}
+2024-03-28T16:23:28 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "GoogleCloudStorage initialized", "timestamp": 1711643008.020333, "uuid": "90ac5738d7dd4b2386d7cebebc9af8db"}
+2024-03-28T16:23:28 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "\u2514\u2500 Loading data frame from bucket 'customer-activity-data-terraform' at key 'transformed_data/time_dimension/part-00000-4d485b3c-b671-4c6f-8624-360d68fda3b3-c000.parquet'...", "timestamp": 1711643008.077176, "uuid": "c8705a8f63604c16b44e081765c3ef19"}
+2024-03-28T16:23:30 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "DONE", "timestamp": 1711643010.019382, "uuid": "aed34bdf559a4751828189363b00eb5c"}
+2024-03-28T16:23:30 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "GoogleCloudStorage initialized", "timestamp": 1711643010.036534, "uuid": "e2f631d323d944f39277e65452cb1730"}
+2024-03-28T16:23:30 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "\u2514\u2500 Loading data frame from bucket 'customer-activity-data-terraform' at key 'transformed_data/fact_order/part-00000-12e06343-69c7-41d9-8107-204da4ceb04e-c000.parquet'...", "timestamp": 1711643010.078758, "uuid": "fd91d9f5222f4ad5b74225f332111b35"}
+2024-03-28T16:23:32 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "DONE", "timestamp": 1711643012.771792, "uuid": "e0336b3661914a28af0a6a5e1502c8bc"}
+2024-03-28T16:23:35 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "--------------------------------------------------------------", "timestamp": 1711643015.813229, "uuid": "8ca04f5588a74b3ea5227bd04f94dcfb"}
+2024-03-28T16:23:35 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "1/1 tests passed.", "timestamp": 1711643015.813671, "uuid": "67abf1600d784348a2ab30d40f8094f1"}
+2024-03-28T16:23:35 {"block_run_id": 1, "block_uuid": "gcs_loader", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_loader", "level": "INFO", "message": "Finish executing block with BlockExecutor.", "timestamp": 1711643015.844769, "uuid": "093dacd6b4864c729dcbcb3c57ac4b0c"}
