+++ docker/mage/mage_data/ashraf-magic/pipelines/batch_pipeline/.logs/1/20240328T162258_317607/gcs_to_bigquery_dim_customer.log
@@ -1,0 +1,9 @@
+2024-03-28T16:23:39 {"block_run_id": 5, "block_uuid": "gcs_to_bigquery_dim_customer", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "Start executing block with BlockExecutor.", "timestamp": 1711643019.356675, "uuid": "ed4751b488874a40a722a5175fec0aca"}
+2024-03-28T16:24:12 {"block_run_id": 5, "block_uuid": "gcs_to_bigquery_dim_customer", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "BigQuery initialized", "timestamp": 1711643052.229209, "uuid": "3a9eeb82ee3448bf86a37f77e437bb4e"}
+2024-03-28T16:24:12 {"block_run_id": 5, "block_uuid": "gcs_to_bigquery_dim_customer", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u2514\u2500 Connecting to BigQuery warehouse...", "timestamp": 1711643052.416769, "uuid": "13c9dcdbeed24a0caa2d6d9f5064b99e"}
+2024-03-28T16:24:12 {"block_run_id": 5, "block_uuid": "gcs_to_bigquery_dim_customer", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "DONE", "timestamp": 1711643052.417634, "uuid": "c55ac41d1f304fbe9f81de1bd441efed"}
+2024-03-28T16:24:12 {"block_run_id": 5, "block_uuid": "gcs_to_bigquery_dim_customer", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "", "timestamp": 1711643052.424592, "uuid": "a618e01fe806404687be04285dd6cec4"}
+2024-03-28T16:24:12 {"block_run_id": 5, "block_uuid": "gcs_to_bigquery_dim_customer", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u251c\u2500", "timestamp": 1711643052.42522, "uuid": "6077bccf0ee1418fa9c9c66b49a15c66"}
+2024-03-28T16:24:12 {"block_run_id": 5, "block_uuid": "gcs_to_bigquery_dim_customer", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u2514\u2500 Exporting data to table 'gothic-sylph-387906.customer_activity_dat.dim_customer'...", "timestamp": 1711643052.425715, "uuid": "8221d422d52a4fd4bce4e384cb6027f9"}
+2024-03-28T16:24:22 {"block_run_id": 5, "block_uuid": "gcs_to_bigquery_dim_customer", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "DONE", "timestamp": 1711643062.379929, "uuid": "729be8eb7ce74e3c812652d3c7590592"}
+2024-03-28T16:24:22 {"block_run_id": 5, "block_uuid": "gcs_to_bigquery_dim_customer", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "Finish executing block with BlockExecutor.", "timestamp": 1711643062.398056, "uuid": "c7003f0ad45940bb91d78e95ad937418"}
