+++ docker/mage/mage_data/ashraf-magic/pipelines/batch_pipeline/.logs/1/20240328T162258_317607/gcs_to_bigquery_dim_delivery_person.log
@@ -1,0 +1,9 @@
+2024-03-28T16:23:39 {"block_run_id": 3, "block_uuid": "gcs_to_bigquery_dim_delivery_person", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "Start executing block with BlockExecutor.", "timestamp": 1711643019.017267, "uuid": "347d0326c0644a009690d73cd1dbce25"}
+2024-03-28T16:24:12 {"block_run_id": 3, "block_uuid": "gcs_to_bigquery_dim_delivery_person", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "BigQuery initialized", "timestamp": 1711643052.238254, "uuid": "2e66bb40e4394e4f8f18974a8f589cd4"}
+2024-03-28T16:24:12 {"block_run_id": 3, "block_uuid": "gcs_to_bigquery_dim_delivery_person", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u2514\u2500 Connecting to BigQuery warehouse...", "timestamp": 1711643052.405703, "uuid": "769062ad049e42998498ac412a848814"}
+2024-03-28T16:24:12 {"block_run_id": 3, "block_uuid": "gcs_to_bigquery_dim_delivery_person", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "DONE", "timestamp": 1711643052.406128, "uuid": "0c3552b7f8814245ba921c2446583954"}
+2024-03-28T16:24:12 {"block_run_id": 3, "block_uuid": "gcs_to_bigquery_dim_delivery_person", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "", "timestamp": 1711643052.508697, "uuid": "6970090380e54618b0fc39da9513d02e"}
+2024-03-28T16:24:12 {"block_run_id": 3, "block_uuid": "gcs_to_bigquery_dim_delivery_person", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u251c\u2500", "timestamp": 1711643052.509124, "uuid": "9e1fc0109b324e169335b6f2514052f9"}
+2024-03-28T16:24:12 {"block_run_id": 3, "block_uuid": "gcs_to_bigquery_dim_delivery_person", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u2514\u2500 Exporting data to table 'gothic-sylph-387906.customer_activity_dat.dim_delivery_person'...", "timestamp": 1711643052.509304, "uuid": "43bc9b5393f947c8b89d7be2f0e21a7b"}
+2024-03-28T16:24:23 {"block_run_id": 3, "block_uuid": "gcs_to_bigquery_dim_delivery_person", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "DONE", "timestamp": 1711643063.031742, "uuid": "a13ff7c879e04c1e9f55798cca414a8d"}
+2024-03-28T16:24:23 {"block_run_id": 3, "block_uuid": "gcs_to_bigquery_dim_delivery_person", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "Finish executing block with BlockExecutor.", "timestamp": 1711643063.09136, "uuid": "6f545d06277b46e0a8afa9a7aa312155"}
