+++ docker/mage/mage_data/ashraf-magic/pipelines/batch_pipeline/.logs/1/20240328T162258_317607/gcs_to_bigquery_dim_location.log
@@ -1,0 +1,9 @@
+2024-03-28T16:23:39 {"block_run_id": 4, "block_uuid": "gcs_to_bigquery_dim_location", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "Start executing block with BlockExecutor.", "timestamp": 1711643019.126546, "uuid": "978fc543a08040eaaad1a0e2a261f17e"}
+2024-03-28T16:24:12 {"block_run_id": 4, "block_uuid": "gcs_to_bigquery_dim_location", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "BigQuery initialized", "timestamp": 1711643052.220156, "uuid": "0c3202f76bfb4234bfd0449215c279b9"}
+2024-03-28T16:24:12 {"block_run_id": 4, "block_uuid": "gcs_to_bigquery_dim_location", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u2514\u2500 Connecting to BigQuery warehouse...", "timestamp": 1711643052.399605, "uuid": "a8d53bb875344f97996634b1b0c2c5cb"}
+2024-03-28T16:24:12 {"block_run_id": 4, "block_uuid": "gcs_to_bigquery_dim_location", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "DONE", "timestamp": 1711643052.400022, "uuid": "b1d47c31e87a46b1aa171b018037997d"}
+2024-03-28T16:24:12 {"block_run_id": 4, "block_uuid": "gcs_to_bigquery_dim_location", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "", "timestamp": 1711643052.56542, "uuid": "156beba8542e4d9f90708deb6266b020"}
+2024-03-28T16:24:12 {"block_run_id": 4, "block_uuid": "gcs_to_bigquery_dim_location", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u251c\u2500", "timestamp": 1711643052.566224, "uuid": "f95ca917bd7b436184784c70f6855564"}
+2024-03-28T16:24:12 {"block_run_id": 4, "block_uuid": "gcs_to_bigquery_dim_location", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u2514\u2500 Exporting data to table 'gothic-sylph-387906.customer_activity_dat.dim_location'...", "timestamp": 1711643052.566777, "uuid": "cac3f7d78d1048fca5dfc090c00f649d"}
+2024-03-28T16:24:22 {"block_run_id": 4, "block_uuid": "gcs_to_bigquery_dim_location", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "DONE", "timestamp": 1711643062.862465, "uuid": "5db120f9d6ae446aaa3ea1e3b937bd6f"}
+2024-03-28T16:24:22 {"block_run_id": 4, "block_uuid": "gcs_to_bigquery_dim_location", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "Finish executing block with BlockExecutor.", "timestamp": 1711643062.876897, "uuid": "afa1a87da03e4062aef61db77417aade"}
