+++ docker/mage/mage_data/ashraf-magic/pipelines/batch_pipeline/.logs/1/20240328T162258_317607/gcs_to_bigquery_dim_time.log
@@ -1,0 +1,9 @@
+2024-03-28T16:23:39 {"block_run_id": 6, "block_uuid": "gcs_to_bigquery_dim_time", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "Start executing block with BlockExecutor.", "timestamp": 1711643019.450276, "uuid": "a28326afa34f4536a66af8ac31a81b2a"}
+2024-03-28T16:24:12 {"block_run_id": 6, "block_uuid": "gcs_to_bigquery_dim_time", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "BigQuery initialized", "timestamp": 1711643052.086847, "uuid": "61f972b4089445a1817e85f9342f88f0"}
+2024-03-28T16:24:12 {"block_run_id": 6, "block_uuid": "gcs_to_bigquery_dim_time", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u2514\u2500 Connecting to BigQuery warehouse...", "timestamp": 1711643052.299464, "uuid": "ce79ec2e6a1747369fa3bdd709c8bddb"}
+2024-03-28T16:24:12 {"block_run_id": 6, "block_uuid": "gcs_to_bigquery_dim_time", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "DONE", "timestamp": 1711643052.300257, "uuid": "4e3f60e9fd47495fa814eba294831a7c"}
+2024-03-28T16:24:12 {"block_run_id": 6, "block_uuid": "gcs_to_bigquery_dim_time", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "", "timestamp": 1711643052.396122, "uuid": "c432100aa74a43188de0b7ab097945dd"}
+2024-03-28T16:24:12 {"block_run_id": 6, "block_uuid": "gcs_to_bigquery_dim_time", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u251c\u2500", "timestamp": 1711643052.39699, "uuid": "978a730466124652b21156a48b65127c"}
+2024-03-28T16:24:12 {"block_run_id": 6, "block_uuid": "gcs_to_bigquery_dim_time", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "\u2514\u2500 Exporting data to table 'gothic-sylph-387906.customer_activity_dat.dim_time'...", "timestamp": 1711643052.397604, "uuid": "b6fc56bfa0d045cc9bc583c92b99ccb1"}
+2024-03-28T16:24:22 {"block_run_id": 6, "block_uuid": "gcs_to_bigquery_dim_time", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "DONE", "timestamp": 1711643062.227641, "uuid": "ef7c03361f4847c7b9b7cda66da3642e"}
+2024-03-28T16:24:22 {"block_run_id": 6, "block_uuid": "gcs_to_bigquery_dim_time", "pipeline_run_id": 1, "pipeline_schedule_id": 1, "pipeline_uuid": "batch_pipeline", "hostname": "bdc66cbcead4", "block_type": "data_exporter", "level": "INFO", "message": "Finish executing block with BlockExecutor.", "timestamp": 1711643062.235314, "uuid": "d82ef85e1fe94a2abd42f344abd3685b"}
