+++ docker/mage/mage_data/ashraf-magic/pipelines/batch_pipeline/.variables/1/20240328T162258_317607/gcs_loader/output_0/type.json
@@ -1,0 +1,1 @@
+{"type": null}
