+++ docker/mage/mage_data/ashraf-magic/pipelines/batch_pipeline/.variables/transformer/output_0/type.json
@@ -1,0 +1,1 @@
+{"type": null}
